requirements.txt
# Core DL
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# HuggingFace ecosystem
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.14.0

# CLIP
openai-clip

# Interpretability
grad-cam>=1.4.8
captum>=0.6.0

# Demo
gradio>=4.0.0

# Data & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.65.0

# Logging (optional)
wandb

# Notebook
ipykernel
ipywidgets
